+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
+## [2.0.0] - 2020-07-21
+
+### Minor Performance Improvements 
+*   internal changes to how cache lookup 
+### Breaking Changes
+*   Rename ObjectMapper.MapProperties to ObjectMapper.Map
+
+
+
+
 ## [1.0.26] - 2019-09-30
 
 ### Bug Fix